+++ src/dbt/include/fabricspark/macros/materializations/models/table/table.sql
@@ -50,6 +50,7 @@
 
   {% do persist_docs(target_relation, model) %}
   {% do persist_constraints(target_relation, model) %}
+  {% do optimize(target_relation) %}
   {{ run_hooks(post_hooks) }}
   {{ return({'relations': [target_relation]})}}